pyproject.toml
[project]
name = "english-evaluation-service"
version = "1.0.0"
description = "Internal English evaluation service using ElevenLabs transcription and Claude analysis"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "PyJWT>=2.9.0",
    "cryptography>=43.0.0",
    "pytz>=2024.1",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "config*", "models*", "schemas*", "routers*", "services*", "repositories*"]
